services/ecological-eval/pyproject.toml
[tool.poetry]
name = "ecological-eval"
version = "0.1.0"
description = "Ecological evaluation service for KindPath"
authors = ["KindPath Foundation"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
kafka-python = "^2.0.2"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"
pydantic = "^2.5.0"
jinja2 = "^3.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"